+++ shm/Fruit.cpp
@@ -30,7 +30,7 @@ size_t Fruit::getPrice() const {
     if (!maxFreshTime_ or !freshTime_) {
         return 0u;
     }
-    const size_t price = std::max(static_cast<size_t>(basePrice_ * freshTime_ / static_cast<double>(maxFreshTime_)),1ul);
+    const size_t price = std::max(static_cast<size_t>(static_cast<double>(basePrice_ * freshTime_) / static_cast<double>(maxFreshTime_)),1ul);
     return price;
 }
